# docs/openapi-generated.yaml
openapi: 3.1.0
info:
  title: 'Bar Assistant API'
...
  contact:
    name: 'Bar Assistant'
    url: 'https://barassistant.app'
    email: info@barassistant.app
  license:
    name: MIT
    url: 'https://github.com/karlomikus/bar-assistant/blob/master/LICENSE'
  version: '{{VERSION}}'
servers:
  -
    url: 'http://localhost:8000/api'
    description: 'Local docker development'
  -
    url: 'https://api.barassistant.app/api'
    description: Production
paths:
...
  /ingredients:
    get:
      tags:
        - Ingredients
      summary: 'List ingredients'
      description: 'Show a list of all ingredients in a bar'
      operationId: listIngredients
      parameters:
        -
          name: Bar-Assistant-Bar-Id
          in: header
          description: 'Database id of a bar. Required if you are not using `bar_id` query string.'
          required: false
          schema:
            type: integer
        -
          name: page
          in: query
          description: 'Set current page number'
          schema:
            type: integer
        -
          name: per_page
          in: query
          description: 'Set number of results per page'
          schema:
            type: integer
        -
          name: filter
          in: query
          description: 'Filter by attributes'
          style: deepObject
          explode: true
          schema:
            properties:
              id:
                type: integer
              name:
                type: string
              name_exact:
                type: string
              category_id:
                type: integer
              origin:
                type: string
              created_user_id:
                type: integer
              on_shopping_list:
                type: boolean
              on_shelf:
                type: boolean
              bar_shelf:
                type: boolean
              strength_min:
                type: number
                format: float
              strength_max:
                type: number
                format: float
              main_ingredients:
                type: string
              complex:
                type: boolean
            type: object
        -
          name: sort
          in: query
          description: 'Sort by attributes. Available attributes: `name`, `created_at`, `strength`, `total_cocktails`.'
          schema:
            type: string
        -
          name: include
          in: query
          description: 'Include additional relationships. Available relations: `parentIngredient`, `varieties`, `prices`, `ingredientParts`, `category`, `images`.'
          schema:
            type: string
      responses:
        '200':
          description: 'Successful response'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Ingredient' } }
                  links: { properties: { first: { type: [string, 'null'] }, last: { type: [string, 'null'] }, prev: { type: [string, 'null'] }, next: { type: [string, 'null'] } }, type: object }
                  meta: { properties: { current_page: { type: integer }, from: { type: integer }, last_page: { type: integer }, links: { type: array, items: { properties: { url: { type: string }, label: { type: string }, active: { type: boolean } }, type: object } }, path: { type: string }, per_page: { type: integer }, to: { type: integer }, total: { type: integer } }, type: object }
                type: object
        '403':
          description: 'You are not authorized for this action.'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/APIError' }
                type: object
    post:
      tags:
        - Ingredients
      summary: 'Create ingredient'
      description: 'Create a new ingredient'
      operationId: saveIngredient
      parameters:
        -
          name: Bar-Assistant-Bar-Id
          in: header
          description: 'Database id of a bar. Required if you are not using `bar_id` query string.'
          required: false
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngredientRequest'
      responses:
        '201':
          description: 'Successful response'
          headers:
            Location:
              description: 'URL of the new resource'
              schema:
                type: string
          content:
            application/json:
              schema:
                required:
                  - data
                properties:
                  data: { $ref: '#/components/schemas/Ingredient' }
                type: object
        '403':
          description: 'You are not authorized for this action.'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/APIError' }
                type: object
...
  '/users/{id}/ingredients':
    get:
      tags:
        - 'Users: Shelf'
      summary: 'List user ingredients'
      description: 'Ingredients that user saved to their shelf'
      operationId: listUserIngredients
      parameters:
        -
          name: id
          in: path
          description: 'Database id of a resource'
          required: true
          schema:
            type: integer
        -
          name: Bar-Assistant-Bar-Id
          in: header
          description: 'Database id of a bar. Required if you are not using `bar_id` query string.'
          required: false
          schema:
            type: integer
        -
          name: page
          in: query
          description: 'Set current page number'
          schema:
            type: integer
        -
          name: per_page
          in: query
          description: 'Set number of results per page'
          schema:
            type: integer
      responses:
        '200':
          description: 'Successful response'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/IngredientBasic' } }
                  links: { properties: { first: { type: [string, 'null'] }, last: { type: [string, 'null'] }, prev: { type: [string, 'null'] }, next: { type: [string, 'null'] } }, type: object }
                  meta: { properties: { current_page: { type: integer }, from: { type: integer }, last_page: { type: integer }, links: { type: array, items: { properties: { url: { type: string }, label: { type: string }, active: { type: boolean } }, type: object } }, path: { type: string }, per_page: { type: integer }, to: { type: integer }, total: { type: integer } }, type: object }
                type: object
  '/users/{id}/cocktails':
    get:
      tags:
        - 'Users: Shelf'
      summary: 'List shelf cocktails'
      description: 'Cocktails that the user can make with ingredients on their shelf'
      operationId: listUserShelfCocktails
      parameters:
        -
          name: id
          in: path
          description: 'Database id of a resource'
          required: true
          schema:
            type: integer
        -
          name: Bar-Assistant-Bar-Id
          in: header
          description: 'Database id of a bar. Required if you are not using `bar_id` query string.'
          required: false
          schema:
            type: integer
        -
          name: page
          in: query
          description: 'Set current page number'
          schema:
            type: integer
        -
          name: per_page
          in: query
          description: 'Set number of results per page'
          schema:
            type: integer
      responses:
        '200':
          description: 'Successful response'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/CocktailBasic' } }
                  links: { properties: { first: { type: [string, 'null'] }, last: { type: [string, 'null'] }, prev: { type: [string, 'null'] }, next: { type: [string, 'null'] } }, type: object }
                  meta: { properties: { current_page: { type: integer }, from: { type: integer }, last_page: { type: integer }, links: { type: array, items: { properties: { url: { type: string }, label: { type: string }, active: { type: boolean } }, type: object } }, path: { type: string }, per_page: { type: integer }, to: { type: integer }, total: { type: integer } }, type: object }
                type: object
  '/users/{id}/cocktails/favorites':
    get:
      tags:
        - 'Users: Shelf'
      summary: 'List favorites'
      description: 'Show a list of cocktails user has favorited'
      operationId: listUserFavoriteCocktails
      parameters:
        -
          name: id
          in: path
          description: 'Database id of a resource'
          required: true
          schema:
            type: integer
        -
          name: Bar-Assistant-Bar-Id
          in: header
          description: 'Database id of a bar. Required if you are not using `bar_id` query string.'
          required: false
          schema:
            type: integer
        -
          name: page
          in: query
          description: 'Set current page number'
          schema:
            type: integer
        -
          name: per_page
          in: query
          description: 'Set number of results per page'
          schema:
            type: integer
      responses:
        '200':
          description: 'Successful response'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/CocktailBasic' } }
                  links: { properties: { first: { type: [string, 'null'] }, last: { type: [string, 'null'] }, prev: { type: [string, 'null'] }, next: { type: [string, 'null'] } }, type: object }
                  meta: { properties: { current_page: { type: integer }, from: { type: integer }, last_page: { type: integer }, links: { type: array, items: { properties: { url: { type: string }, label: { type: string }, active: { type: boolean } }, type: object } }, path: { type: string }, per_page: { type: integer }, to: { type: integer }, total: { type: integer } }, type: object }
                type: object
  '/users/{id}/ingredients/batch-store':
    post:
      tags:
        - 'Users: Shelf'
      summary: 'Save user ingredients'
      description: 'Save multiple ingredients to user shelf'
      operationId: batchStoreUserIngredients
      parameters:
        -
          name: id
          in: path
          description: 'Database id of a resource'
          required: true
          schema:
            type: integer
        -
          name: Bar-Assistant-Bar-Id
          in: header
          description: 'Database id of a bar. Required if you are not using `bar_id` query string.'
          required: false
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                ingredients:
                  type: array
                  items: { type: integer }
              type: object
      responses:
        '204':
          description: 'Successful response'
        '403':
          description: 'You are not authorized for this action.'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/APIError' }
                type: object
        '404':
          description: 'Resource record not found.'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/APIError' }
                type: object
  '/users/{id}/ingredients/batch-delete':
    post:
      tags:
        - 'Users: Shelf'
      summary: 'Delete user ingredients'
      description: 'Delete multiple ingredients from user shelf'
      operationId: batchDeleteUserIngredients
      parameters:
        -
          name: id
          in: path
          description: 'Database id of a resource'
          required: true
          schema:
            type: integer
        -
          name: Bar-Assistant-Bar-Id
          in: header
          description: 'Database id of a bar. Required if you are not using `bar_id` query string.'
          required: false
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                ingredients:
                  type: array
                  items: { type: integer }
              type: object
      responses:
        '204':
          description: 'Successful response'
        '403':
          description: 'You are not authorized for this action.'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/APIError' }
                type: object
        '404':
          description: 'Resource record not found.'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/APIError' }
                type: object
...
components:
  schemas:
...
    APIError:
      required:
        - type
        - message
      properties:
        type:
          type: string
          example: api_error
        message:
          type: string
          example: 'Resource record not found.'
      type: object
...
    Ingredient:
      required:
        - id
        - slug
        - name
        - description
        - origin
        - color
        - created_at
        - updated_at
        - strength
      properties:
        id:
          type: integer
          example: 1
        slug:
          type: string
          example: gin-1
        name:
          type: string
          example: Gin
        strength:
          type: number
          format: float
          example: 40
        description:
          type:
            - string
            - 'null'
          example: 'Gin is a type of alcoholic spirit'
        origin:
          type:
            - string
            - 'null'
          example: Worldwide
        created_at:
          type: string
          format: date-time
        updated_at:
          type:
            - string
            - 'null'
          format: date-time
        images:
          type: array
          items:
            $ref: '#/components/schemas/Image'
        parent_ingredient:
          oneOf:
            -
              $ref: '#/components/schemas/IngredientBasic'
            -
              type: 'null'
        color:
          type: string
          example: '#ffffff'
        category:
          oneOf:
            -
              $ref: '#/components/schemas/IngredientCategory'
            -
              type: 'null'
        cocktails_count:
          type: integer
          example: 12
        varieties:
          type: array
          items:
            $ref: '#/components/schemas/IngredientBasic'
        cocktails:
          type: array
          items:
            properties:
              id:
                type: integer
                example: 1
              slug:
                type: string
                example: old-fashioned-1
              name:
                type: string
                example: 'Old fashioned'
            type: object
        created_user:
          $ref: '#/components/schemas/UserBasic'
        updated_user:
          oneOf:
            -
              $ref: '#/components/schemas/UserBasic'
            -
              type: 'null'
        access:
          type: array
          items:
            properties:
              can_edit:
                type: boolean
                example: true
              can_delete:
                type: boolean
                example: true
            type: object
        ingredient_parts:
          type: array
          items:
            $ref: '#/components/schemas/IngredientBasic'
        prices:
          type: array
          items:
            $ref: '#/components/schemas/IngredientPrice'
        in_shelf:
          type: boolean
        in_bar_shelf:
          type: boolean
        in_shopping_list:
          type: boolean
        used_as_substitute_for:
          type: array
          items:
            $ref: '#/components/schemas/IngredientBasic'
        can_be_substituted_with:
          type: array
          items:
            $ref: '#/components/schemas/IngredientBasic'
      type: object
...
  securitySchemes:
    user_token:
      type: http
      scheme: bearer
security:
  -
    user_token: []
...